# [CUDA] Parallelize upsampling OPS across the batch/channel dimension.