--- a/test/test_nn.py
+++ b/test/test_nn.py
@@ -9614,7 +9614,11 @@ class TestNNDeviceType(NNTestCase):
 
     @parametrize_test("memory_format", [torch.contiguous_format, torch.channels_last])
     def test_upsamplingBilinear2d_aa_correctness(self, device, memory_format):
-        t_in = torch.arange(3 * 8 * 8, dtype=torch.float, device=device).reshape(1, 3, 8, 8)
+        # NOTE: We expand the batch dim such that `b*c` is above the maximum
+        # size of CUDA grid z-dimension (2**16)
+        shape = [23000, 3, 8, 8]
+        t_in = torch.arange(3 * 8 * 8, dtype=torch.float, device=device).reshape(1, *shape[1:])
+        t_in = t_in.expand(shape)
         t_in = t_in.contiguous(memory_format=memory_format)
         # This expected result is obtain using PIL.Image.resize
         # for c in range(3):
@@ -9626,7 +9630,7 @@ class TestNNDeviceType(NNTestCase):
             106.75, 109.96428, 145.0357, 148.25, 170.75, 173.9643
         ], device=device, dtype=t_in.dtype).reshape(1, 3, 2, 2)
         t_out = F.interpolate(t_in, size=(2, 2), mode="bilinear", align_corners=False, antialias=True)
-        self.assertEqual(expected_out, t_out)
+        self.assertEqual(expected_out.expand([*shape[:2], 2, 2]), t_out)
 
     @parametrize_test("memory_format", [torch.contiguous_format, torch.channels_last])
     @parametrize_test("mode", ["bilinear", "bicubic"])